Fix undefined behavior in SuperFastHash
Cherry-pick from chromium:
https://crrev.com/c/3717978/5/base/third_party/superfasthash/superfasthash.c

BUG=b:240230674
TEST=None

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4131096
Tested-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>

+++ cras/src/common/sfh.c
@@ -55,7 +55,7 @@ uint32_t SuperFastHash(const char *data, int len, uint32_t hash)
 	/* Main loop */
 	for (; len > 0; len--) {
 		hash += get16bits(data);
-		tmp = (get16bits(data + 2) << 11) ^ hash;
+		tmp = (uint32_t)(get16bits(data + 2) << 11) ^ hash;
 		hash = (hash << 16) ^ tmp;
 		data += 2 * sizeof(uint16_t);
 		hash += hash >> 11;
@@ -66,7 +66,7 @@ uint32_t SuperFastHash(const char *data, int len, uint32_t hash)
 	case 3:
 		hash += get16bits(data);
 		hash ^= hash << 16;
-		hash ^= ((signed char)data[sizeof(uint16_t)]) << 18;
+		hash ^= (uint32_t)(((signed char)data[sizeof(uint16_t)]) << 18);
 		hash += hash >> 11;
 		break;
 	case 2:
@@ -75,7 +75,7 @@ uint32_t SuperFastHash(const char *data, int len, uint32_t hash)
 		hash += hash >> 17;
 		break;
 	case 1:
-		hash += (signed char)*data;
+		hash += (uint32_t)((signed char)*data);
 		hash ^= hash << 10;
 		hash += hash >> 1;
 	}
